OneDrive/Desktop/PandaBrief/requirements.txt
google-generativeai
suno-api
python-dotenv
